+++ misc/prelink.js
@@ -6,7 +6,7 @@ console.log(`> linking all peer dependencies of ${pkg.name} as children to be pr
 
 Object.keys(pkg.peerDependencies).map(dep => {
     if (dep.indexOf('nexrender') !== -1) {
-        const command = `npm link --no-package-lock --legacy-peer-deps ../${dep.replace('@nexrender/', 'nexrender-')}`
+        const command = `npm link --no-package-lock --legacy-peer-deps ../${dep.replace('@create-global/nexrender-', 'nexrender-')}`
         console.log(`executing > ${command}`)
         execSync(command)
     }