+++ packages/nexrender-worker/src/bin.js
@@ -2,7 +2,7 @@
 
 const arg       = require('arg')
 const chalk     = require('chalk')
-const { init } = require('@nexrender/core')
+const { init } = require('@create-global/nexrender-core')
 const {start}   = require('./index')
 const {version} = require('../package.json')
 const rimraf    = require('rimraf')